git-svn-id: svn+ssh://combine-101/home/svn_bio/shrimp/branches/region-filter@720 450597fd-3c47-0410-84d8-f85f2794411d

+++ common/fasta.c
@@ -410,7 +410,7 @@ fasta_get_next_read_with_range(fasta_t fasta, read_entry * re )
 			return (false);
 		}
 		re->qual = (char *)xmalloc(quality_length + 17);
-		for (i=0; i<quality_length; i++) {
+		for (i=0; i<(int)quality_length; i++) {
 			re->qual[i]=MAX((char)fasta->parse_buffer[i],'!');
 		}	
 		//memcpy(re->qual, fasta->parse_buffer, quality_length);
